refactor(ax-alloc): remove ax-allocator dependency, simplify to TLSF/buddy-slab backends
Remove the two-level allocation architecture (byte allocator + bitmap
page allocator from ax-allocator). TLSF now uses rlsf directly as a
single-level allocator. buddy-slab remains unchanged.

- Remove ax-allocator dependency, add rlsf directly
- Remove slab/buddy/page-alloc-4g/page-alloc-64g features
- tlsf and buddy-slab are mutually exclusive, neither enabled by default
- Add build.rs for cfg(tlsf)/cfg(buddy_slab) generation
- Add stub_impl.rs with unimplemented!() when no backend is selected
- Clean up downstream Cargo.toml feature propagation
