+++ os/arceos/modules/axalloc/src/lib.rs
@@ -7,6 +7,7 @@
 
 #![no_std]
 
+#[allow(unused_imports)]
 #[macro_use]
 extern crate log;
 extern crate alloc;
@@ -52,10 +53,12 @@ impl Usages {
         Self([0; UsageKind::VARIANTS.len()])
     }
 
+    #[allow(dead_code)]
     fn alloc(&mut self, kind: UsageKind, size: usize) {
         self.0[kind as usize] += size;
     }
 
+    #[allow(dead_code)]
     fn dealloc(&mut self, kind: UsageKind, size: usize) {
         self.0[kind as usize] -= size;
     }
@@ -175,19 +178,23 @@ pub trait AllocatorOps {
     fn usages(&self) -> Usages;
 }
 
-// Select implementation based on features.
-#[cfg(feature = "buddy-slab")]
+// Select implementation based on build.rs-generated cfg flags.
+#[cfg(buddy_slab)]
 mod buddy_slab;
-#[cfg(feature = "buddy-slab")]
-use buddy_slab as imp;
+#[cfg(not(any(tlsf, buddy_slab)))]
+mod stub_impl;
+#[cfg(tlsf)]
+mod tlsf_impl;
 
-#[cfg(not(feature = "buddy-slab"))]
-mod default_impl;
-#[cfg(not(feature = "buddy-slab"))]
-use default_impl as imp;
-#[cfg(feature = "buddy-slab")]
+#[cfg(buddy_slab)]
+use buddy_slab as imp;
+#[cfg(buddy_slab)]
 pub use imp::init_percpu_slab;
 pub use imp::{DefaultByteAllocator, GlobalAllocator, global_add_memory, global_init};
+#[cfg(not(any(tlsf, buddy_slab)))]
+use stub_impl as imp;
+#[cfg(tlsf)]
+use tlsf_impl as imp;
 
 /// Returns the reference to the global allocator.
 pub fn global_allocator() -> &'static GlobalAllocator {
