+++ os/arceos/ulib/arceos-rust/lib/Cargo.toml
@@ -42,12 +42,8 @@ myplat = ["ax-feat/myplat"]
 
 # Memory
 alloc = ["ax-api/alloc", "ax-feat/alloc", "ax-io/alloc"]
-alloc-buddy = ["ax-feat/alloc-buddy"]
-alloc-slab = ["ax-feat/alloc-slab"]
 alloc-tlsf = ["ax-feat/alloc-tlsf"]
 dma = ["ax-api/dma", "ax-feat/dma"]
-page-alloc-4g = ["ax-feat/page-alloc-4g"] # Support up to 4G memory capacity
-page-alloc-64g = ["ax-feat/page-alloc-64g"] # Support up to 64G memory capacity
 paging = ["ax-feat/paging"]
 tls = ["ax-feat/tls"]
